pyproject.toml
[project]
name = "telecardio"
version = "0.1.0"
description = "Distribute holters between doctors and display statistics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml==6.0.2",
]
